thermal_anomaly_india_usa_project/airflow/requirements.txt
apache-airflow==2.10.3
apache-airflow-providers-google==10.26.0
google-cloud-storage==2.19.0
google-cloud-bigquery==3.25.0
pandas==2.1.4
pyarrow==17.0.0
pyspark==3.5.2
